State management implemenatation
Installed the providers package
Implementation of state management
Refactor -> Appending the Screen in file and class name for screen
