+++ lib/main.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
-import './screens/product_detail.dart';
-import './screens/product_overview.dart';
+import 'package:provider/provider.dart';
+import './screens/product_detail_screen.dart';
+import './screens/product_overview_screen.dart';
+import './providers/products.dart';
 
 void main() => runApp(PettyShopApp());
 
@@ -8,18 +10,21 @@ class PettyShopApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Petty Shop',
-      theme: ThemeData(
-        primarySwatch: Colors.purple,
-        accentColor: Colors.deepOrangeAccent,
-        fontFamily: 'Lato',
-        splashColor: Colors.black87
+    return ChangeNotifierProvider(
+      builder: (context) => Products(),
+      child: MaterialApp(
+        title: 'Petty Shop',
+        theme: ThemeData(
+          primarySwatch: Colors.purple,
+          accentColor: Colors.deepOrangeAccent,
+          fontFamily: 'Lato',
+          splashColor: Colors.black87
+        ),
+        home: ProductOverviewScreen(),
+        routes: {
+          ProductDetailScreen.routeName: (ctx) => ProductDetailScreen()
+        },
       ),
-      home: ProductOverview(),
-      routes: {
-        ProductDetail.routeName: (ctx) => ProductDetail()
-      },
     );
   }
 }
