+++ lib/providers/products.dart
@@ -1,0 +1,127 @@
+import 'package:flutter/material.dart';
+import './../models/product.dart';
+
+class Products with ChangeNotifier {
+  List<Product> _items = [
+    Product(
+      id: '1',
+      title: 'Onions',
+      description: 'About the onion',
+      price: '123.00',
+      image: 'https://images-na.ssl-images-amazon.com/images/I/71rAIKxakYL._SX679_.jpg'
+    ),
+    Product(
+      id: '2',
+      title: 'Tomato',
+      description: 'About the tomato',
+      price: '43.00',
+      image: 'https://dam.farmjournal.com/m/16f06f5f2c6978c5/webimage-946FE331-C128-4F61-AC32684C11417AC3.png'
+    ),
+    Product(
+      id: '3',
+      title: 'Dhall',
+      description: 'Dhall',
+      price: '28.50',
+      image: 'https://images-na.ssl-images-amazon.com/images/I/81f17QBywiL._SY550_.jpg'
+    ),
+    Product(
+      id: '4',
+      title: 'Ragi Flour',
+      description: 'About Ragi Flour',
+      price: '24.30',
+      image: 'http://timbaktu-organic.org/wp-content/uploads/2015/03/Malted-Ragi-Flour-600-600x600.jpg'
+    ),
+    Product(
+      id: '5',
+      title: 'Ladies Finger',
+      description: 'About Ragi Flour',
+      price: '12.30',
+      image: 'https://5.imimg.com/data5/BR/ND/MY-1768518/fresh-lady-finger-500x500.jpg'
+    ),
+    Product(
+      id: '7',
+      title: 'Cotton Candy',
+      description: 'About the onion',
+      price: '2.00',
+      image: 'https://odditymall.com/includes/content/move-over-cheese-balls-cotton-candy-balls-are-here-to-take-over-thumb.jpg'
+    ),
+    Product(
+      id: '8',
+      title: 'Dairy Milk',
+      description: 'About the tomato',
+      price: '15.00',
+      image: 'https://www.waangoo.com/content/images/thumbs/0012491_cadbury-dairy-milk_600.jpeg'
+    ),
+    Product(
+      id: '8',
+      title: 'Murukku',
+      description: 'Dhall',
+      price: '0.50',
+      image: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQo7PdnetDJurTlKQFRmFzpuLXCKdT0oDd0fhOkA2RDGJQQe3ks&s'
+    ),
+    Product(
+      id: '9',
+      title: 'Banana',
+      description: 'About Ragi Flour',
+      price: '2',
+      image: 'https://api.time.com/wp-content/uploads/2019/11/gettyimages-459761948.jpg?w=600&quality=85'
+    ),
+    Product(
+      id: '10',
+      title: 'Hamam Soap',
+      description: 'About Ragi Flour',
+      price: '30.00',
+      image: 'https://www.waangoo.com/content/images/thumbs/0009971_hamam-soap-100gm_600.jpeg'
+    ),
+    Product(
+      id: '11',
+      title: 'Cinthol',
+      description: 'About the onion',
+      price: '40.00',
+      image: 'https://images-na.ssl-images-amazon.com/images/I/81aVEokflXL._SL1500_.jpg'
+    ),
+    Product(
+      id: '12',
+      title: 'Brinjal',
+      description: 'About the tomato',
+      price: '43.00',
+      image: 'https://5.imimg.com/data5/JD/VD/PH/ANDROID-51655394/product-jpeg-500x500.jpg'
+    ),
+    Product(
+      id: '13',
+      title: 'Sugar',
+      description: 'Sugar',
+      price: '38.00',
+      image: 'http://worldwidegeneral.com/wp-content/uploads/2017/08/20161023-Udhsugar-417x561.jpg'
+    ),
+    Product(
+      id: '14',
+      title: 'Parachute',
+      description: 'About Ragi Flour',
+      price: '35.00',
+      image: 'https://images-na.ssl-images-amazon.com/images/I/61D7y8hUGmL._SL1500_.jpg'
+    ),
+    Product(
+      id: '15',
+      title: 'Peanut',
+      description: 'About Ragi Flour',
+      price: '4.00',
+      image: 'https://3.imimg.com/data3/XS/RQ/MY-11974068/java-groundnut-500x500.jpg'
+    ),
+  ];
+
+  List<Product> get items {
+    return [..._items];
+  }
+
+  void addProduct() {
+    // TODO: Add the product here
+    notifyListeners();
+  }
+
+  Product findById(String id) {
+    return _items.firstWhere((product) {
+      return product.id == id;
+    });
+  }
+}
