+++ lib/screens/product_detail_screen.dart
@@ -1,14 +1,17 @@
 import 'package:flutter/material.dart';
-import './../models/product.dart';
-import './../DATA.dart';
+import 'package:provider/provider.dart';
+import './../providers/products.dart';
 
-class ProductDetail extends StatelessWidget {
+class ProductDetailScreen extends StatelessWidget {
 
   static const routeName = '/product-detail';
   @override
   Widget build(BuildContext context) {
     final String productId = ModalRoute.of(context).settings.arguments as String;
-    final Product product = PRODUCTS.firstWhere((prod) => prod.id == productId);
+    final product = Provider.of<Products>(
+      context,
+      listen: false
+    ).findById(productId);
     return Scaffold(
       appBar: AppBar(
         title: Text(
