+++ pubspec.yaml
@@ -27,6 +27,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  provider: ^3.0.0
 
 
 # For information on the generic Dart part of this file, see the
